Reverted deletion of the ExceptionExtensions class in Api Extensions and marked th eonly method there as Obsolete instead to prevent breaking those who are using its. Addressed some other comments from PR.
