+++ src/Microsoft.Health.Fhir.Core/Extensions/ExceptionExtensions.cs
@@ -40,19 +40,5 @@ namespace Microsoft.Health.Fhir.Core.Extensions
         {
             return e is RequestEntityTooLargeException || e?.InnerException is RequestEntityTooLargeException;
         }
-
-        public static Exception GetInnerMostException(this Exception input)
-        {
-            EnsureArg.IsNotNull(input, nameof(input));
-
-            Exception current = input;
-
-            while (current.InnerException != null)
-            {
-                current = current.InnerException;
-            }
-
-            return current;
-        }
     }
 }
